+++ packages/analytics/dashboard-renderer/src/components/GoldenSignalsRenderer.vue
@@ -68,17 +68,19 @@ const options = computed<ProviderProps>(() => ({
   :deep(.kong-ui-public-metric-card-container) {
     height: 100%;
 
-    @media (min-width: ($kui-breakpoint-phablet - 1px)) {
-      height: 80%;
-    }
-
     .metricscard {
-      justify-content: space-evenly;
-
       @media (min-width: ($kui-breakpoint-phablet - 1px)) {
         &:not(:last-of-type) {
           border-right: $kui-border-width-10 solid $kui-color-border;
         }
+      }
+
+      // These overrides are here because, for now, we only want the bolder metric titles
+      // on dashboards.  In other places, the golden signal cards tend to be embedded into KCards which
+      // already have titles.
+      .metricscard-title.lg {
+        font-size: $kui-font-size-40;
+        font-weight: $kui-font-weight-semibold;
       }
     }
   }