+++ packages/analytics/analytics-chart/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-chart@2.4.1...@kong-ui-public/analytics-chart@2.4.2) (2024-03-08)
+
+
+### Bug Fixes
+
+* replace CopyUuid with KCopy [khcp-9465] ([#1210](https://github.com/Kong/public-ui-components/issues/1210)) ([e409d60](https://github.com/Kong/public-ui-components/commit/e409d60b7ab483c200ed6bb33f43ff571d6e6b88))
+
+
+
+
+
+## [2.4.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-chart@2.4.0...@kong-ui-public/analytics-chart@2.4.1) (2024-03-07)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-chart
+
+
+
+
+
+# [2.4.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-chart@2.3.4...@kong-ui-public/analytics-chart@2.4.0) (2024-03-07)
+
+
+### Features
+
+* **dashboard-renderer:** expand component test [MA-2629] ([#1193](https://github.com/Kong/public-ui-components/issues/1193)) ([9aaf215](https://github.com/Kong/public-ui-components/commit/9aaf215449105147f6ff35405f181da533105474))
+
+
+
+
+
+## [2.3.4](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-chart@2.3.3...@kong-ui-public/analytics-chart@2.3.4) (2024-03-07)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-chart
+
+
+
+
+
+## [2.3.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-chart@2.3.2...@kong-ui-public/analytics-chart@2.3.3) (2024-03-06)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-chart
+
+
+
+
+
 ## [2.3.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-chart@2.3.1...@kong-ui-public/analytics-chart@2.3.2) (2024-03-04)