+++ packages/analytics/metric-cards/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.41](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/metric-cards@0.6.40...@kong-ui-public/metric-cards@0.6.41) (2024-03-08)
+
+
+### Bug Fixes
+
+* replace CopyUuid with KCopy [khcp-9465] ([#1210](https://github.com/Kong/public-ui-components/issues/1210)) ([e409d60](https://github.com/Kong/public-ui-components/commit/e409d60b7ab483c200ed6bb33f43ff571d6e6b88))
+
+
+
+
+
+## [0.6.40](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/metric-cards@0.6.39...@kong-ui-public/metric-cards@0.6.40) (2024-03-07)
+
+**Note:** Version bump only for package @kong-ui-public/metric-cards
+
+
+
+
+
+## [0.6.39](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/metric-cards@0.6.38...@kong-ui-public/metric-cards@0.6.39) (2024-03-06)
+
+
+### Bug Fixes
+
+* **analytics:** update metric card styles [MA-2658] ([#1206](https://github.com/Kong/public-ui-components/issues/1206)) ([d507993](https://github.com/Kong/public-ui-components/commit/d507993cb18362c94032df129147a254f9b70e0d))
+
+
+
+
+
 ## [0.6.38](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/metric-cards@0.6.37...@kong-ui-public/metric-cards@0.6.38) (2024-02-27)