+++ packages/core/app-layout/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.12](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/app-layout@4.0.11...@kong-ui-public/app-layout@4.0.12) (2024-03-08)
+
+
+### Bug Fixes
+
+* replace CopyUuid with KCopy [khcp-9465] ([#1210](https://github.com/Kong/public-ui-components/issues/1210)) ([e409d60](https://github.com/Kong/public-ui-components/commit/e409d60b7ab483c200ed6bb33f43ff571d6e6b88))
+
+
+
+
+
+## [4.0.11](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/app-layout@4.0.10...@kong-ui-public/app-layout@4.0.11) (2024-03-07)
+
+**Note:** Version bump only for package @kong-ui-public/app-layout
+
+
+
+
+
+## [4.0.10](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/app-layout@4.0.9...@kong-ui-public/app-layout@4.0.10) (2024-03-06)
+
+**Note:** Version bump only for package @kong-ui-public/app-layout
+
+
+
+
+
 ## [4.0.9](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/app-layout@4.0.8...@kong-ui-public/app-layout@4.0.9) (2024-02-27)