+++ packages/entities/entities-gateway-services/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.5.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-gateway-services@2.5.2...@kong-ui-public/entities-gateway-services@2.5.3) (2024-03-08)
+
+
+### Bug Fixes
+
+* replace CopyUuid with KCopy [khcp-9465] ([#1210](https://github.com/Kong/public-ui-components/issues/1210)) ([e409d60](https://github.com/Kong/public-ui-components/commit/e409d60b7ab483c200ed6bb33f43ff571d6e6b88))
+
+
+
+
+
+## [2.5.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-gateway-services@2.5.1...@kong-ui-public/entities-gateway-services@2.5.2) (2024-03-07)
+
+**Note:** Version bump only for package @kong-ui-public/entities-gateway-services
+
+
+
+
+
+## [2.5.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-gateway-services@2.5.0...@kong-ui-public/entities-gateway-services@2.5.1) (2024-03-06)
+
+**Note:** Version bump only for package @kong-ui-public/entities-gateway-services
+
+
+
+
+
+# [2.5.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-gateway-services@2.4.32...@kong-ui-public/entities-gateway-services@2.5.0) (2024-03-06)
+
+
+### Features
+
+* **entities-gateway-services:** support control plane column ([#1212](https://github.com/Kong/public-ui-components/issues/1212)) ([a796ec1](https://github.com/Kong/public-ui-components/commit/a796ec1bfe544ad5fed37a77b8dd19fc6711fa37))
+
+
+
+
+
+## [2.4.32](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-gateway-services@2.4.31...@kong-ui-public/entities-gateway-services@2.4.32) (2024-03-05)
+
+**Note:** Version bump only for package @kong-ui-public/entities-gateway-services
+
+
+
+
+
 ## [2.4.31](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-gateway-services@2.4.30...@kong-ui-public/entities-gateway-services@2.4.31) (2024-03-01)
 
 **Note:** Version bump only for package @kong-ui-public/entities-gateway-services